// class-components/src/types/types.ts
export interface Character {
  name: string;
  gender: string;
  height: string;
  eye_color: string;
  homeworld: string;
  url: string;
}

export interface DetailedCharacter {
  name: string;
  birth_year: string;
  gender: string;
  height: string;
  mass: string;
  eye_color: string;
  skin_color: string;
  homeworld: string;
  url: string;
}